+++ connet/src/main/java/houseInception/connet/contoller/PrivateRoomController.java
@@ -1,6 +1,7 @@
 package houseInception.connet.contoller;
 
 import houseInception.connet.dto.*;
+import houseInception.connet.dto.privateRoom.*;
 import houseInception.connet.response.BaseResponse;
 import houseInception.connet.response.BaseResultDto;
 import houseInception.connet.service.PrivateRoomService;
