+++ connet/src/main/java/houseInception/connet/dto/friend/FriendFilterDto.java
@@ -1,4 +1,4 @@
-package houseInception.connet.dto;
+package houseInception.connet.dto.friend;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
