+++ connet/src/main/java/houseInception/connet/dto/friend/FriendType.java
@@ -1,4 +1,4 @@
-package houseInception.connet.dto;
+package houseInception.connet.dto.friend;
 
 public enum FriendType {
     FRIEND, NONE, REQUEST, WAIT
