+++ connet/src/main/java/houseInception/connet/dto/login/TokenResDto.java
@@ -1,4 +1,4 @@
-package houseInception.connet.dto;
+package houseInception.connet.dto.login;
 
 import lombok.Getter;
