+++ connet/src/main/java/houseInception/connet/repository/FriendCustomRepository.java
@@ -2,7 +2,7 @@ package houseInception.connet.repository;
 
 import houseInception.connet.dto.ActiveUserResDto;
 import houseInception.connet.dto.DefaultUserResDto;
-import houseInception.connet.dto.FriendFilterDto;
+import houseInception.connet.dto.friend.FriendFilterDto;
 
 import java.util.List;
