+++ connet/src/main/java/houseInception/connet/service/GptRoomService.java
@@ -5,11 +5,11 @@ import houseInception.connet.domain.gptRoom.GptRoom;
 import houseInception.connet.domain.gptRoom.GptRoomChat;
 import houseInception.connet.domain.gptRoom.GptRoomUser;
 import houseInception.connet.dto.*;
+import houseInception.connet.dto.GptRoom.*;
 import houseInception.connet.exception.GptRoomException;
 import houseInception.connet.externalServiceProvider.gpt.GptApiProvider;
 import houseInception.connet.externalServiceProvider.gpt.GptResDto;
 import houseInception.connet.repository.GptRoomRepository;
-import houseInception.connet.repository.UserRepository;
 import houseInception.connet.service.util.DomainValidatorUtil;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
