+++ connet/src/main/java/houseInception/connet/service/PrivateRoomService.java
@@ -6,11 +6,11 @@ import houseInception.connet.domain.privateRoom.PrivateChat;
 import houseInception.connet.domain.privateRoom.PrivateRoom;
 import houseInception.connet.domain.privateRoom.PrivateRoomUser;
 import houseInception.connet.dto.*;
+import houseInception.connet.dto.privateRoom.*;
 import houseInception.connet.exception.PrivateRoomException;
 import houseInception.connet.externalServiceProvider.gpt.GptApiProvider;
 import houseInception.connet.externalServiceProvider.s3.S3ServiceProvider;
 import houseInception.connet.repository.PrivateRoomRepository;
-import houseInception.connet.repository.UserBlockRepository;
 import houseInception.connet.service.util.DomainValidatorUtil;
 import houseInception.connet.socketManager.SocketServiceProvider;
 import houseInception.connet.socketManager.dto.PrivateChatSocketDto;
