+++ connet/src/test/java/houseInception/connet/service/PrivateRoomServiceTest.java
@@ -4,8 +4,8 @@ import houseInception.connet.domain.*;
 import houseInception.connet.domain.privateRoom.PrivateChat;
 import houseInception.connet.domain.privateRoom.PrivateRoom;
 import houseInception.connet.domain.privateRoom.PrivateRoomUser;
-import houseInception.connet.dto.*;
 import houseInception.connet.dto.chatEmoji.ChatEmojiResDto;
+import houseInception.connet.dto.privateRoom.*;
 import houseInception.connet.exception.PrivateRoomException;
 import houseInception.connet.exception.UserBlockException;
 import houseInception.connet.repository.PrivateRoomRepository;
